requirements.txt
yfinance
langchain_pinecone
openai
python-dotenv
langchain_huggingface
sentence_transformers
python-dotenv